Setting subdomain in capybara tests
